# Conflicts:
#	src/distributed_legder/src/blockchain/block.rs
#	src/distributed_legder/src/main.rs

+++ src/distributed_legder/src/blockchain/transaction.rs
@@ -1,8 +1,14 @@
 use serde::{Serialize, Deserialize};
 
+/// Represents a transaction in the blockchain.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Transaction {
+    /// The sender of the transaction.
     pub sender: String,
+
+    /// The recipient of the transaction.
     pub recipient: String,
-    pub amount: f32,
+
+    /// The amount of the transaction.
+    pub amount: f64,
 }
